+++ Example/cauli-ios-example/cauli-ios-example/AppDelegate.swift
@@ -7,7 +7,7 @@
 //
 
 import UIKit
-import Cauli
+import Cauliframework
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {