+++ Pods/Target Support Files/Pods-CauliTests/Pods-CauliTests-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Pods_CauliTests : NSObject
-@end
-@implementation PodsDummy_Pods_CauliTests
-@end